% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/geohost.R
\name{geohost}
\alias{geohost}
\title{IP address lookup}
\usage{
geohost(ip = "", position = TRUE)
}
\arguments{
\item{ip}{a character vector specifying a ip (e.g., "12.215.42.19").
The default value is no ip is specified and the host ip is used.}

\item{position}{whether or not output lat/lng infomation. The default value is TRUE.}
}
\value{
a vector with information of contry_name, country_code, city, and ip.
If numerous ips are inputted, a data.frame is returned.
}
\description{
geocodes an IP address using HostIP.info API \url{http://www.hostip.info}.
}
\examples{
\dontrun{
# geocode host IP
geohost()
# no lat/lng infomation
geohost(position = F)
# specify an IP for geocoding
geohost(ip = "12.215.42.19")
geohost(ip = "12.215.42.19", position = F)
# geocode multiple IPs
geohost(ip = c("61.135.169.81", "12.215.42.19"))
geohost(ip = c("61.135.169.81", "12.215.42.19"), position = F)
}
}
\author{
Jun Cai (\email{cai-j12@mails.tsinghua.edu.cn}), PhD student from
Center for Earth System Science, Tsinghua University
}
\seealso{
\code{\link{geocode}}, \code{\link{revgeocode}}.

HostIP.info IP Address Lookup API at \url{http://www.hostip.info/use.html}
}

