% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/revgeocode.R
\name{revgeocode}
\alias{revgeocode}
\title{Reverse geocode}
\usage{
revgeocode(latlng, ics = c("WGS-84", "GCJ-02", "BD-09"), api = c("google",
  "baidu"), key = "", output = c("address", "addressc"),
  messaging = FALSE, time = 0)
}
\arguments{
\item{latlng}{a location in latitude/longitude format}

\item{ics}{the coordinate system of inputing location, including WGS-84, GCJ-02 
and BD-09, which are the GCSs of Google Earth, Google Map in China and Baidu 
Map, respectively. For location out of China, ics is automatically set to 'WGS-84'
and other values are ignored.}

\item{api}{use Google or Baidu Maps Geocoding API}

\item{key}{an API key must be provided when calling Baidu Maps Geocoding API. When calling Google Maps Geocoding API, you'd better provide an API key, though it is not necessary.}

\item{output}{formatted address or formmatted address with address components}

\item{messaging}{turn messaging on/off. The default value is FALSE.}

\item{time}{the time interval to revgeocode, in seconds. Default value is zero. 
When you revgeocode multiple locations, set a proper time interval to avoid 
exceeding usage limits. For details see 
\url{https://developers.google.com/maps/documentation/geocoding/usage-limits}}
}
\value{
a data.frame with variables address or detail address components
}
\description{
reverse geocodes a lat/lng location using Google or Baidu Maps Geocoding API.  Note that in 
most cases by using this function you are agreeing to the Google Maps API Terms 
of Service at \url{https://developers.google.com/maps/terms} or the Baidu Maps 
API Terms of Use at \url{http://developer.baidu.com/map/law.htm}.
}
\details{
note that the google maps api limits to 2,500 free requests per day and 50 requests per second.
}
\examples{
\dontrun{
# reverse geocode Beijing Railway Station
revgeocode(c(39.90358, 116.421), ics = 'WGS-84', api = 'google', 
           output = 'address')
revgeocode(c(39.90498, 116.4272), ics = 'GCJ-02', api = 'google', 
           output = 'address', messaging = TRUE)
revgeocode(c(39.91103, 116.4337), ics = 'BD-09', api = 'google', 
           output = 'addressc')
revgeocode(c(39.91103, 116.4337), ics = 'BD-09', api = 'baidu', 
           key = 'your baidu maps api key', output = 'address')
revgeocode(c(39.90498, 116.4272), ics = 'GCJ-02', api = 'baidu', 
           key = 'your baidu maps api key', output = 'address', messaging = TRUE)
revgeocode(c(39.90358, 116.421), ics = 'WGS-84', api = 'baidu', 
           key = 'your baidu maps api key', output = 'addressc')
# reverse geocode multiple locations
latlng = data.frame(lat = c(39.99837, 39.98565), lng = c(116.3203, 116.2998))
revgeocode(latlng, ics = 'WGS-84', api = 'google', output = 'address')
revgeocode(latlng, ics = 'WGS-84', api = 'google', output = 'address', time = 2)
}
}
\author{
Jun Cai (\email{cai-j12@mails.tsinghua.edu.cn}), PhD candidate from 
Department of Earth System Science, Tsinghua University
}
\seealso{
\code{\link{geocode}}, \code{\link{geohost}}.

Google Maps API at \url{https://developers.google.com/maps/documentation/geocoding/} 
and Baidu Maps API at \url{http://lbsyun.baidu.com/index.php?title=webapi/guide/webservice-geocoding}
}

